requirements.txt
discord.py==1.6.0
discord-py-slash-command==3.0.1
async-timeout==3.0.1
